// lib/types/schemas.backend.ts
const chatHistoryMessageSchema = z.object({
    message: z.string().nullable(),
    type: z.string().nullable(),
    agent_avatar: z.string().nullable(),
    agent_id: z.number().nullable(),
    agent_name: z.string().nullable(),
    api_called: z.boolean().nullable(),
    chatbot_id: z.string().nullable(),
    created_at: z.string(),
    debug_json: z.any().nullable(),
    extra_params: z.any(),
    from_user: z.boolean().nullable(),
    handoff_happened_during_office_hours: z.boolean().nullable(),
    id: z.number(),
    knowledgebase_called: z.boolean().nullable(),
    session_id: z.string(),
    updated_at: z.string().nullable(),
})
// ...
type ChatHistoryMessageType = z.infer<typeof chatHistoryMessageSchema>;